+++ core/classes/Templates/AssetResolver.php
@@ -27,8 +27,9 @@ class AssetResolver extends AssetTree {
         }
 
         foreach ($assets as $asset) {
-            $this->validateAsset($asset);
-            $this->_assets[$asset] = parent::ASSET_TREE[$asset];
+            if ($this->validateAsset($asset)) {
+                $this->_assets[$asset] = parent::ASSET_TREE[$asset];
+            }
         }
     }
 
@@ -42,7 +43,7 @@ class AssetResolver extends AssetTree {
         $css = [];
         $js = [];
 
-        $assets = Util::determineOrder(array_map(function($key, $value) {
+        $assets = Util::determineOrder(array_map(static function(string $key, array $value) {
             return [
                 'name' => $key,
                 'after' => $value['depends'] ?? $value['after'] ?? [],
@@ -99,6 +100,10 @@ class AssetResolver extends AssetTree {
             foreach ($asset['js'] as $jsFile) {
                 $js[] = $this->buildPath($jsFile, 'js');
             }
+        }
+
+        if (!in_array($name, $this->_assets)) {
+            $this->_assets[$name] = $asset;
         }
     }